lites-compat.h: declare malloc as unsigned int, matching LITES
The LITES server links. The emulator did not build because of a wrong
declaration here.

emulator/e_mach_msg_server.c declares "extern void *malloc(unsigned
int)", which is correct for i386 where size_t is unsigned int. This
header declared unsigned long, so the variadic macro below rewrote that
declaration into a conflicting one and GCC rejected it.

For the record, the server binary built at this point is
startup.Lites.1.1.u3.STD+WS+osfmach3, 953,360 bytes stripped and
1,065,424 unstripped: a complete BSD UNIX personality linked against
OSFMK 7.3.

The link also needs -z muldefs. LITES is a kernel and defines its own
printf, vsprintf, sprintf and sleep; libsa_mach provides standalone
versions, and its printf.o gets pulled in for another symbol, bringing
the collisions with it. LITES's definitions come first on the command
line and are the ones wanted.

+++ tools/lites/lites-compat.h
@@ -37,7 +37,13 @@ typedef mig_reply_error_t mig_reply_header_t;
  * Both arities work: malloc(n) and malloc(n, M_RTABLE, M_DONTWAIT) both
  * reach the one-argument allocator.
  */
-extern void *malloc(unsigned long);
+/*
+ * i386: size_t is unsigned int. LITES declares malloc as
+ * void *malloc(unsigned int) in emulator/e_mach_msg_server.c, so this
+ * must agree or the macro below rewrites that declaration into a
+ * conflicting one.
+ */
+extern void *malloc(unsigned int);
 extern void  free(void *);
 
 #define malloc(sz, ...)  (malloc)(sz)
